Add OutputPayload class for carrying presentation logic
